% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_vcf.R
\name{as.winfun}
\alias{as.winfun}
\alias{region}
\alias{samples}
\title{Window Extractor Functions}
\usage{
as.winfun(f, max.n, samples, region)

region(f)

samples(f)
}
\arguments{
\item{f}{A function (for as.winfun), or a window extractor function (class \code{winfun}).}

\item{max.n}{Index of the largest window.}

\item{samples}{Character vector of sample IDs corresponding to columns of extracted data.}

\item{region}{A function taking an integer vector returning the chromosome, start, and end of the corresponding windows.}
}
\description{
A window extractor function ("winfun") takes an integer vector and returns an integer matrix, 
giving the number of alternate alleles for each sample for the corresponding windows.
Such functions also have three attributes: \code{max.n}, giving the index of the largest window,
\code{samples}, the sample IDs that are extracted (corresponding to the columns of the matrix that is returned),
and \code{region}, which is a function that takes an integer vector and returns a data frame giving chromosome, start, and end of the corresponding windows.
}
\section{Functions}{
\itemize{
\item \code{region}: Returns the \code{region} function of a window extractor function.

\item \code{samples}: Gives the sample IDs returned by a window extractor function.
}}

