% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_vcf.R
\name{dim.winfun}
\alias{dim.winfun}
\title{Gives the number of samples of matrices returned by a window extractor function (access with ncol( )).}
\usage{
\method{dim}{winfun}(f)
}
\description{
Gives the number of samples of matrices returned by a window extractor function (access with ncol( )).
}

