% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_windows.R
\name{eigen_windows}
\alias{eigen_windows}
\title{Find Eigenvectors and Eigenvalues for Covariance Matrices in Windows}
\usage{
eigen_windows(data, k, do.windows = NULL, win, mc.cores = 1, ...)
}
\arguments{
\item{data}{Either a numeric matrix or a window extractor function, that takes a positive integer \code{n} and returns the numeric matrix that correspond to the \code{n}th window.}

\item{k}{Number of eigenvalue/eigenvector pairs to return.}

\item{do.windows}{Vector of integers that correspond to the windows used.}

\item{win}{If \code{data} is a matrix, each contiguous block of \code{win} rows of the matrix \code{data} will be one window.}

\item{mc.cores}{If this is greater than 1, parallel::mclapply will be used.}

\item{...}{Other parameters to be passed to \code{win.fn}.}
}
\value{
A numeric matrix with one row for each window; the first k columns giving eigenvalues and each subsequent group of ncol(data) columns giving the corresponding eigenvector.
The result has an attribute, "npc", which is equal to the number of eigenvalues computed (k).
}
\description{
Subtracts row means from the input matrix, then
divides the rows into windows of a given size,
computes the covariance matrix for each,
and returns the given number of eigenvector/eigenvalue pairs.
Omits the last (short) window.
}

