% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_windows.R
\name{eigen_windows_matrix}
\alias{eigen_windows_matrix}
\title{Sets up eigen_windows_winfn to work on a matrix.}
\usage{
eigen_windows_matrix(data, k, win, w, do.windows = NULL, mc.cores)
}
\description{
Sets up eigen_windows_winfn to work on a matrix.
}

