% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enclosing_circle.R
\name{enclosing_circle}
\alias{enclosing_circle}
\alias{plot_circle}
\title{Minimum Enclosing Circle}
\usage{
enclosing_circle(xy)

plot_circle(circ, n = 200, plot.it = TRUE, ...)
}
\arguments{
\item{xy}{The 2D coordinates.}

\item{circ}{A circle as returned by \code{enclosing_circle}.}

\item{n}{The number of angles to draw the circle at.}

\item{plot.it}{Actually do plotting or just return points?}
}
\value{
A named list with entries \code{ctr} (the coordinates of the center), \code{rad} (the radius),
\code{three} (the three points in xy that define the circle),
and \code{index} (the indices of those three points in xy).
}
\description{
Finds the minimum enclosing circle to a set of points,
"Skyum algorithm based on the convex hull", as implemented at 
   http://www.uni-kiel.de/psychologie/rexrepos/posts/diagBounding.html#minimum-enclosing-circle
}
\section{Functions}{
\itemize{
\item \code{plot_circle}: Plots a circle.
}}

