% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_vcf.R
\name{enclosing_region}
\alias{enclosing_region}
\title{Find Minimal Enclosing Regions}
\usage{
enclosing_region(sites)
}
\arguments{
\item{sites}{Named list of positions, as output by \code{vcf_positions()}.}
}
\value{
A data frame with variables chrom, start, and end.
}
\description{
Given a named list of positions, whose names correspond to chromosomes,
return the data frame of regions with names chrom,start,end that encloses them.
}

