% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enclosing_circle.R
\name{isBiggerThan90}
\alias{isBiggerThan90}
\title{Check if the angle at B in triangle ABC exceeds 90 degrees}
\usage{
isBiggerThan90(xy)
}
\description{
Check if the angle at B in triangle ABC exceeds 90 degrees
}

