% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_vcf.R
\name{multi_vcf_query}
\alias{multi_vcf_query}
\title{Query Multiple VCF Files}
\usage{
multi_vcf_query(chrom.list, file = names(chrom.list), regions, ...)
}
\arguments{
\item{chrom.list}{A list of character vectors of chromosome names, the \code{n}th saying which chromosomes are available in the \code{n}th vcf file.}

\item{file}{The name(s) of the indexed vcf file(s).}

\item{regions}{A data frame containing chrom, start, and end of the regions to be extracted.}

\item{...}{Other arguments passed to \code{vcf_query}.}
}
\value{
The results of the separate \code{vcf_query} calls, \code{rbind}'ed together
(this will *not* necessarily be in the same order as \code{regions}!).
}
\description{
Applies \code{vcf_query} to multiple VCF files: must specify which chromosomes are in which file with \code{chrom.list}.
}

