% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_vcf.R
\name{multi_vcf_query_fn}
\alias{multi_vcf_query_fn}
\title{Function to Query Multiple VCF Files}
\usage{
multi_vcf_query_fn(chrom.list, file = names(chrom.list), regions, ...)
}
\arguments{
\item{chrom.list}{A list of character vectors of chromosome names, the \code{n}th saying which chromosomes are available in the \code{n}th vcf file.}

\item{file}{The name(s) of the indexed vcf file(s).}

\item{regions}{A data frame containing chrom, start, and end of the regions to be extracted.}

\item{...}{Other arguments passed to \code{vcf_query}.}
}
\value{
A function, say, \code{f}, so that \code{f(n)} is the result of \code{n}th \code{vcf_query} call.
}
\description{
Returns a function that when called with an integer \code{n} will return the \code{n}th region.
}

