% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_vcf.R
\name{query_genotypes}
\alias{query_genotypes}
\title{Query a VCF File For Numeric Genotypes}
\usage{
query_genotypes(file, regions, samples, verbose = FALSE)
}
\arguments{
\item{file}{The name of the indexed vcf file.}

\item{regions}{A data frame containing chrom, start, and end of the regions to be extracted.}

\item{samples}{A character vector of sample names to be extracted.}

\item{verbose}{Whether to output with \code{cat()} the bcftools calls.}
}
\value{
An integer matrix, giving the number of alternate alleles for each sample.
}
\description{
Uses bcftools to query vcf or bcf files for genotypes in a collection of regions,
returning the genotypes in numeric format.
Assumes the data are diploid.
}

