% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{recode_numeric}
\alias{recode_numeric}
\title{Convert Matrix of Alleles to Numeric}
\usage{
recode_numeric(x, ploidy = 2, triallelic = TRUE, alleles = c("A", "C",
  "G", "T"))
}
\arguments{
\item{x}{Character matrix of alleles.}

\item{ploidy}{Number of columns per individual.}

\item{triallelic}{Include triallelic sites? Defaults to TRUE; otherwise sets these to NA.}

\item{alleles}{A character vector of valid alleles.}
}
\value{
An integer matrix with one row per site and one column per individual.
}
\description{
Recodes a character matrix, where each row is a site and each \code{ploidy} columns are the alleles of an individual, into numeric format,
counting the number of major alleles, where "major" is the most common allele at that site, and all others are lumped as "minor".
}
\details{
For speed, currently only deals with alleles that are A, C, G, or T (not indels).

Values range from 0 (all minor alleles) to \code{ploidy} (all major alleles).
}

