% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_pca.R
\name{running_cov}
\alias{running_cov}
\title{Construct the Covariance Matrix from the Data in Pieces}
\usage{
running_cov(f, n)
}
\arguments{
\item{f}{A function such that f(k) returns the k-th chunk of data.}

\item{n}{The number of chunks of data, or a vector of indices.}
}
\value{
A numeric matrix:
If a and b are vectors, then
   mean(a[1:(n+m)]) == (n/(n+m)) * mean(a[1:n]) + (1/(n+m)) * sum(a[(n+1):(n+m)])
and
   cov(a,b) = sum(a*b)/(n-1) - sum(a)*sum(b)/(n*(n-1))
where the sums are over the n shared nonmissings.
}
\description{
This will compute the covariance matrix for a dataset
that is only available in disjoint pieces (because it is too big to keep in memory at once, for instance).
}

