% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_vcf.R
\name{vcf_positions}
\alias{vcf_positions}
\title{Read Position From a VCF File}
\usage{
vcf_positions(file)
}
\arguments{
\item{file}{The name of the indexed vcf files.}
}
\value{
A named list of integer vectors; names correspond to chromosomes.
}
\description{
Reads the CHROM and POS columns of a VCF file, with bcftools,
returning a list of positions named by the corresponding chromosomes.
}

