% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_windows.R
\name{eigen_windows_winfn}
\alias{eigen_windows_winfn}
\title{Does the work of eigen_windows.}
\usage{
eigen_windows_winfn(win.fn, do.windows = NULL, k, w, mc.cores = 1, ...)
}
\description{
Does the work of eigen_windows.
}

