% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_vcf.R
\name{vcf_windower}
\alias{vcf_windower}
\title{Construct a Window Extractor for a VCF File}
\usage{
vcf_windower(file, size, type, sites = vcf_positions(file),
  samples = vcf_samples(file))
}
\arguments{
\item{file}{The name of the indexed vcf file.}

\item{size}{The size of the window.}

\item{type}{The units of the window: 'bp' or 'snp'?}

\item{sites}{The positions in the VCF file, as returned by \code{vcf_positions}.}

\item{samples}{A character vector of sample names to be extracted.}
}
\value{
A class "winfun" window extractor function that returns an integer matrix, giving the number of alternate alleles for each sample.
Such functions also have three attributes: \code{max.n}, giving the index of the largest window,
\code{samples}, the sample IDs that are extracted (corresponding to the columns of the matrix that is returned),
and \code{region}, which is a function that takes an integer vector and returns a data frame giving chromosome, start, and end of the corresponding windows.
}
\description{
Returns a function that, given an integer \code{n}, returns a numeric matrix of genotypes from the the \code{n}th window of the VCF file.
}

